管理制度

+++ src/views/quality/newPage/managedObject.vue
@@ -80,13 +80,13 @@
 										</div>
 										<div class="block-item">
 											<div class="block-item-label">登记部门<i class="require-icon"></i></div>
-											<div class="block-item-value">指挥部</div>
+											<div class="block-item-value">{{formData.groupName}}</div>
 										</div>
 									</div>
 									<div class="block-line">
 										<div class="block-item">
 											<div class="block-item-label">登记人<i class="require-icon"></i></div>
-											<div class="block-item-value">卢益辉</div>
+											<div class="block-item-value">{{formData.userName}}</div>
 										</div>
 										<div class="block-item">
 											<div class="block-item-label">发布时间</div>
@@ -113,26 +113,27 @@
 										<span style="font-size: 12px;margin-left: 40px;" v-show="!previewMode">支持上传 jpg/jpeg png mp4 docx doc xlsx xls pdf 文件，且不超过 50M</span>
 									</div>
 									<div class="block-line">
-										<el-button size="small" type="primary" v-show="!previewMode">点击上传</el-button>
+										<upload v-show="!previewMode" @afterUp="afterUpAttach($event)"></upload>
+										<!-- <el-button size="small" type="primary" v-show="!previewMode">点击上传</el-button> -->
 										<el-button size="small" type="primary" v-show="previewMode">下载全部</el-button>
 									</div>
 									<div class="block-table">
-										<el-table :data="annexTableData" style="width: 100%" border
+										<el-table :data="formData.attachment" style="width: 100%" border
 											class="have_scrolling">
 											<el-table-column type="index" width="50" align="center" label="序号">
 											</el-table-column>
-											<el-table-column prop="pro" align="center" label="附件" show-overflow-tooltip>
+											<el-table-column prop="fileName" align="center" label="附件" show-overflow-tooltip>
 											</el-table-column>
-											<el-table-column prop="qualityfirstname" width="160px" align="center"
+											<el-table-column prop="uploadTime" width="160px" align="center"
 												label="上传日期">
 											</el-table-column>
-											<el-table-column prop="qualitysecondname" width="120px" align="center"
+											<el-table-column prop="creatorName" width="120px" align="center"
 												label="上传人">
 											</el-table-column>
 											<el-table-column fixed="right" width="120" align="center" label="操作">
 												<template slot-scope="{ row, $index }">
-													<el-button type="primary" size="mini">下载</el-button>
-													<el-button type="danger" size="mini">预览</el-button>
+													<el-button type="text" size="mini" @click="downLoadFile(row.fileUrl)">下载</el-button>
+													<el-button type="text" size="mini" @click="viewImg(row)">预览</el-button>
 												</template>
 											</el-table-column>
 										</el-table>
@@ -167,7 +168,7 @@
 									<el-collapse-item name="1">
 										<template slot="title">
 											<span class="process-index">1</span>
-											 吕冬敏(2022-02-22 16:45:11) 
+											 吕冬敏({{formData.createTime}}) 
 										</template>
 										<div class="log-line">
 											<div class="log-line-label">项目名称：</div>
@@ -184,29 +185,29 @@
 										</div>
 										<div class="log-line">
 											<div class="log-line-label">发布时间：</div>
-											<div class="log-line-value">2022年02月22日</div>
+											<div class="log-line-value">{{formData.publishDate}}</div>
 										</div>
 										<div class="log-line">
 											<div class="log-line-label">标题：</div>
-											<div class="log-line-value">诸暨235国道质量管理体系</div>
+											<div class="log-line-value">{{formData.title}}</div>
 										</div>
 										<div class="log-line">
 											<div class="log-line-label">上传</div>
 										</div>
-										<el-table :data="annexTableData" style="width: 100%" border
+										<el-table :data="formData.attachment" style="width: 100%" border
 											class="have_scrolling">
-											<el-table-column prop="pro" align="center" label="附件" show-overflow-tooltip>
+											<el-table-column type="fileName" align="center" label="附件" show-overflow-tooltip>
 											</el-table-column>
-											<el-table-column prop="qualityfirstname" width="160px" align="center"
+											<el-table-column prop="uploadTime" width="160px" align="center"
 												label="上传日期">
 											</el-table-column>
-											<el-table-column prop="qualitysecondname" width="120px" align="center"
+											<el-table-column prop="creatorName" width="120px" align="center"
 												label="上传人">
 											</el-table-column>
 											<el-table-column fixed="right" width="80" align="center" label="操作">
 												<template slot-scope="{ row, $index }">
-													<el-button type="primary" size="mini">下载</el-button>
-													<el-button type="danger" size="mini">预览</el-button>
+													<el-button type="text" size="mini" @click="downLoadFile(row.fileUrl)">下载</el-button>
+													<el-button type="text" size="mini" @click="viewImg(row)">预览</el-button>
 												</template>
 											</el-table-column>
 										</el-table>
@@ -217,6 +218,12 @@
 					</div>
 				</el-aside>
 			</el-container>
+		</el-dialog>
+		<el-dialog title="预览" :visible.sync="viewImgVisible" :fullscreen="true" width="100%" height="100%">
+			{{viewImgUrl}}
+			<img v-if="viewImgType=='img'" style="width: 100%; height: 100%" :src="viewImgUrl"/>
+			<!-- <el-image style="width: 100%; height: 100%" :src="viewImgUrl" :fit="fill"></el-image> -->
+			<iframe width="100%" height="100%" style="min-height:800px" v-if="viewImgType=='pdf'" :src="viewImgUrl" frameborder="0"></iframe>
 		</el-dialog>
 	</el-container>
 </template>
@@ -243,6 +250,10 @@
 	return fmt;
 	}
 	import * as api from "@/api/quality";
+	import store from "@/store/index";
+	import upload from "../../common/upload.vue"
+	import { getUserInfo } from "@/api/user";
+	import { downLoadFile } from "@/utils/download";
 	export default {
 		data() {
 			return {
@@ -261,34 +272,31 @@
 					totalPage:1,
 					pageSize: 10,
 				},
+				userInfo: {},
 				formData: {
 					id: '',
 					title: '测试', // 标题
-					publishDate: '2022-05-06T16:00:00', // 发布时间
-					projectId: '0', // 项目id
-									// 登记人
-									// 登记部门
-									// 项目名称 
+					publishDate: '2022-05-06', // 发布时间
+					projectId: '1', // 项目id
+					userName: '',	// 登记人
+					groupName: '',	// 登记部门
+					projectName: '',// 项目名称 
 				    attachment: [ // 附件
-						// {
-						// "createTime": 0,
-						// "createUid": "",
-						// "creatorName": "",
-						// "fileName": "",
-						// "filePath": "",
-						// "fileSize": 0,
-						// "fileUrl": "",
-						// "id": ""
-						// }
 					]
-				}
+				},
+				viewImgVisible: false,
+				viewImgType: 'img',
+				viewImgUrl: '',
 			};
 		},
 		created() {},
-		components: {},
+		components: {
+			upload
+		},
 		computed: {},
 		mounted() {
 			this.query();
+			this.getUserInfo();
 		},
 		methods: {
 			query() {
@@ -297,34 +305,107 @@
 				});
 				
 			},
+			getUserInfo() {
+				getUserInfo(localStorage.getItem('ID')).then(res => {
+					this.userInfo = res.data.userInfo;
+				});
+			},
 			addNew() {
+				this.previewMode = false;
 				this.dialogFormVisible = true;
+				const hasGetUserInfo = store.getters.name;
+				
+				this.formData.userName = hasGetUserInfo;
+				this.formData.groupName = this.userInfo.GROUPNAME;
+				this.formData.id = '';
+				this.formData.attachment = [];
 			},
 			addManagedObject() {
 				if (this.formData.publishDate instanceof Date) this.formData.publishDate = this.formData.publishDate.format("yyyy-MM-dd");
 				api.addOrUpdateManagementObjectList(this.formData).then((res) => {
 				  	this.query();
 					this.dialogFormVisible = false;
+
 				});
 			},
 			deleteManagementObject(id) {
 				// /web/api/v1/manageTarget/id
-				api.deleteManagementObject(id).then((res) => {
-				  	this.query();
+				this.$confirm('确认是否删除?', '提示', {
+					confirmButtonText: '确定',
+					cancelButtonText: '取消',
+					type: 'warning'
+				}).then(() => {
+					api.deleteManagementObject(id).then((res) => {
+						this.query();
+						this.$message({
+							type: 'success',
+							message: '删除成功!'
+						});
+					}).catch(error => {
+						this.$message({
+							type: 'fail',
+							message: '删除失败!'
+						});
+					});
 				});
 			},
 			viewDetail(row) {
-				this.formData = row;
+				const that = this;
+				// this.formData = row;
 				this.dialogTitle = "查看详情";
 				this.dialogFormVisible = true;
 				this.previewMode = true;
+				api.getManagementObject(row.id).then((res) => {
+				  	if (res && res.data) {
+						for (let i = 0; i < res.data.attachment.length; i++) {
+							const item = res.data.attachment[i];
+							item.uploadTime = new Date(item.uploadTime).format('yyyy-MM-dd hh:mm:ss')
+						}
+						this.formData = res.data;
+						getUserInfo(res.data.createUserId).then(res => {
+							that.formData.userName = res.data.userInfo.NAME;
+							that.formData.groupName = this.userInfo.GROUPNAME;
+							that.$forceUpdate();
+						});
+					}
+				});
+			},
+			viewImg(row) {
+				// store.getters.lookUrl
+				console.log(row)
+				let format = '';
+				if (row.fileName.indexOf('.png') > -1) format = 'img';
+				if (row.fileName.indexOf('.pdf') > -1) format = 'pdf';
+				this.viewImgType = format;
+				this.viewImgUrl = store.getters.lookUrl + row.fileUrl;
+				this.viewImgVisible = true;
+			},
+			downLoadFile(url) {
+				downLoadFile(url)
 			},
 			handleSizeChange(val) {
 				console.log(`每页 ${val} 条`);
 			},
 			handleCurrentChange(val) {
 				console.log(`当前页: ${val}`);
-			}
+			},
+			afterUpAttach(data){
+				api.getFileInfo({id: data['fileId']}).then((res) => {
+				  	if (res) {
+						debugger
+					}
+				});
+				this.formData.attachment.push({
+					uploadTime: new Date(data['uploadTime']).format('yyyy-MM-dd hh:mm:ss'),
+					fileName: data['fileName'],
+					fileUrl: data['fileId'],
+					creatorName: this.userInfo.NAME,
+					"createUid": "",
+					"filePath": "",
+					"fileSize": 0,
+					"id": ""
+				})
+			},
 		},
 	};
 </script>
